refactor(fsd): Step 4 — features layer migration
Move auth, voc-list-filter, voc-create, voc-status-change, notification,
master-cache into features/ with backward-compat re-export stubs on old paths.
