+++ frontend/src/features/auth/model/useAuth.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+import { AuthContext } from './AuthContext';
+import type { AuthContextValue } from './AuthContext';
+
+export function useAuth(): AuthContextValue {
+  const ctx = useContext(AuthContext);
+  if (!ctx) throw new Error('useAuth must be used within AuthProvider');
+  return ctx;
+}
