+++ frontend/src/features/notification/index.ts
@@ -1,0 +1,3 @@
+export { NotificationProvider, NotificationContext } from './model/NotificationContext';
+export type { AppNotification, NotificationContextValue } from './model/NotificationContext';
+export { useNotifications } from './model/useNotifications';
