+++ frontend/src/features/voc-create/index.ts
@@ -1,0 +1,1 @@
+export { useCreateVoc } from './model/useCreateVoc';
