+++ frontend/src/features/voc-list-filter/ui/vocGridLayout.ts
@@ -1,0 +1,4 @@
+// Shared grid template for VOC list rows (header + data rows must align).
+// 7 columns: expand chevron, issue_code, title (flex), status, assignee, priority, created_at.
+export const VOC_GRID_COLUMNS = '22px 144px 1fr 115px 108px 84px 96px';
+export const VOC_GRID_PADDING_X = 'var(--sp-5)';
