+++ frontend/src/features/voc-status-change/index.ts
@@ -1,0 +1,1 @@
+export { useChangeVocStatus } from './model/useChangeVocStatus';
