+++ frontend/src/features/voc/components/VocReviewDrawer.tsx
@@ -4,8 +4,8 @@ import { Dialog, DialogContent, DialogDescription, DialogTitle } from '@shared/u
 import { cn } from '@shared/lib/cn';
 import { vocApi } from '../../../api/voc';
 import { queryKeys } from '../../../api/queryKeys';
-import { useRole } from '../../../hooks/useRole';
-import { AuthContext } from '../../../contexts/AuthContext';
+import { useRole } from '@features/auth/model/useRole';
+import { AuthContext } from '@features/auth/model/AuthContext';
 import { type InternalNote } from '../../../../../shared/contracts/voc';
 import { VocPermissionGate } from '../../../components/voc/VocPermissionGate';
 import { LoadingState } from '@shared/ui/skeleton';
