+++ frontend/src/features/voc/useVocList.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { vocApi } from '../../api/voc';
 import { queryKeys } from '../../api/queryKeys';
-import { useRole } from '../../hooks/useRole';
+import { useRole } from '@features/auth/model/useRole';
 import type { VocFilter, VocListQuery } from '../../../../shared/contracts/voc';
 
 export function useVocList(
