+++ frontend/src/hooks/useAuth.ts
@@ -1,8 +1,2 @@
-import { useContext } from 'react';
-import { AuthContext, AuthContextValue } from '../contexts/AuthContext';
-
-export function useAuth(): AuthContextValue {
-  const ctx = useContext(AuthContext);
-  if (!ctx) throw new Error('useAuth must be used within AuthProvider');
-  return ctx;
-}
+// Moved to features/auth/model/useAuth (Step 4)
+export { useAuth } from '@features/auth/model/useAuth';
