+++ frontend/src/hooks/useMasterCache.ts
@@ -1,8 +1,2 @@
-import { useContext } from 'react';
-import { MasterCacheContext, MasterCacheContextValue } from '../contexts/MasterCacheContext';
-
-export function useMasterCache(): MasterCacheContextValue {
-  const ctx = useContext(MasterCacheContext);
-  if (!ctx) throw new Error('useMasterCache must be used within MasterCacheProvider');
-  return ctx;
-}
+// Moved to features/master-cache/model/useMasterCache (Step 4)
+export { useMasterCache } from '@features/master-cache/model/useMasterCache';
