+++ frontend/src/hooks/useNotifications.ts
@@ -1,8 +1,2 @@
-import { useContext } from 'react';
-import { NotificationContext, NotificationContextValue } from '../contexts/NotificationContext';
-
-export function useNotifications(): NotificationContextValue {
-  const ctx = useContext(NotificationContext);
-  if (!ctx) throw new Error('useNotifications must be used within NotificationProvider');
-  return ctx;
-}
+// Moved to features/notification/model/useNotifications (Step 4)
+export { useNotifications } from '@features/notification/model/useNotifications';
